requirements.txt
loguru==0.7.2
playwright==1.48.0
PyExecJS==1.5.1
Requests==2.32.3
